moved logic to hook from component
