+++ src/components/LaunchedEscrows/LaunchedEscrows.tsx
@@ -1,48 +1,26 @@
 import * as React from 'react';
-import { useQuery } from '@apollo/client';
 import { Box, Button } from '@mui/material';
 import { SkeletonLaunched } from 'src/components';
-import { LAUNCHED_ESCROWS } from 'src/queries';
 import { InfiniteScroll } from 'src/components/Scroll';
 import { LaunchedEscrowsView } from './LaunchedEscrowsView';
+import { useLaunchedEscrowsHook } from './hooks';
 
 export const LaunchedEscrows: React.FC = (): React.ReactElement => {
-  const [page, setPage] = React.useState(0);
-
-  const { loading, error, data, refetch } = useQuery(LAUNCHED_ESCROWS, {
-    variables: { skip: page },
-  });
-
-  React.useEffect(() => {
-    refetch({ skip: page });
-  }, [page, refetch]);
-
-  const launchedEscrows = React.useMemo(
-    () => data?.launchedEscrows || [],
-    [data]
-  );
-  const [escrows, setEscrows] = React.useState(launchedEscrows || []);
-
-  React.useEffect(() => {
-    setEscrows((prev: []) => [...prev, ...launchedEscrows]);
-  }, [launchedEscrows]);
-
-  const [onCalled, setOnCalled] = React.useState(false);
-
-  const loadMoreNumbers = () => {
-    setPage((p) => p + 10);
-    setOnCalled((prev) => !prev);
-  };
-
-  const hanldeCall = () => {
-    setOnCalled((prev) => !prev);
-  };
+  const {
+    escrows,
+    loading,
+    error,
+    data,
+    handlePress,
+    loadMoreNumbers,
+    onPressed,
+  } = useLaunchedEscrowsHook();
 
   return (
     <Box>
       <InfiniteScroll
         onBottomHit={loadMoreNumbers}
-        onCalled={onCalled}
+        onPressed={onPressed}
         isLoading={loading}
       >
         <Box
@@ -62,11 +40,10 @@ export const LaunchedEscrows: React.FC = (): React.ReactElement => {
           display: 'flex',
           alignItems: 'center',
           alignContent: 'center',
-          // width: '100%',
           margin: '20px auto 40px',
         }}
       >
-        <Button variant="outlined" onClick={hanldeCall} disabled={loading}>
+        <Button variant="outlined" onClick={handlePress} disabled={loading}>
           Load more Escrows
         </Button>
       </Box>
